+++ src/utils/__test__/add-to-score.test.js
