+++ src/utils/__test__/check-easy-equations.test.js
