+++ src/utils/__test__/check-equation.test.js
