+++ src/utils/__test__/simplify-equation-numbers.test.js
