+++ src/utils/__test__/validate-operators.test.js
