+++ src/utils/add-to-score.js
