+++ src/utils/check-easy-equations.js
