+++ src/utils/check-equation.js
