+++ src/utils/check-used-patterns.js
