+++ src/utils/simplify-equation-numbers.js
