+++ src/utils/validate-operators.js
